Consolidate modifiers types
