--- src/@types/held-modifier-config.ts
@@ -1,8 +1,0 @@
-import type { PokemonHeldItemModifier } from "#modifier/modifier";
-import type { PokemonHeldItemModifierType } from "#modifier/modifier-type";
-
-export interface HeldModifierConfig {
-  modifier: PokemonHeldItemModifierType | PokemonHeldItemModifier;
-  stackCount?: number;
-  isTransferable?: boolean;
-}
