--- src/@types/modifier-pool.ts
@@ -1,5 +1,0 @@
-import type { WeightedModifierType } from "#modifier/modifier-type";
-
-export interface ModifierPool {
-  [tier: string]: WeightedModifierType[];
-}
