--- src/@types/modifier-predicate.ts
@@ -1,3 +1,0 @@
-import type { Modifier } from "#modifier/modifier";
-
-export type ModifierPredicate<T extends Modifier = Modifier> = (modifier: T) => boolean;
