+++ src/battle-scene.ts
@@ -118,8 +118,7 @@ import { settings } from "#system/settings-manager";
 import type TrainerData from "#system/trainer-data";
 import { type Voucher, vouchers } from "#system/voucher";
 import { allTrainerConfigs } from "#trainer-configs/all-trainer-configs";
-import type { HeldModifierConfig } from "#types/held-modifier-config";
-import type { ModifierPredicate } from "#types/modifier-predicate";
+import type { HeldModifierConfig, ModifierPredicate } from "#types/modifiers-types";
 import type { PokemonSpeciesFilter } from "#types/pokemon-species-filter";
 import type { AnySettingKey, SettingsUpdateEventArgs } from "#types/settings";
 import type { AbstractConstructor } from "#types/utility-types";
