+++ src/data/mystery-encounters/encounters/absolute-avarice-encounter.ts
@@ -35,7 +35,7 @@ import type MysteryEncounter from "#mystery-encounters/mystery-encounter";
 import { MysteryEncounterBuilder } from "#mystery-encounters/mystery-encounter";
 import { MysteryEncounterOptionBuilder } from "#mystery-encounters/mystery-encounter-option";
 import { PersistentModifierRequirement } from "#mystery-encounters/mystery-encounter-requirements";
-import type { HeldModifierConfig } from "#types/held-modifier-config";
+import type { HeldModifierConfig } from "#types/modifiers-types";
 import { getPokemonSpecies } from "#utils/pokemon-utils";
 import { randInt } from "#utils/random-utils";
 import i18next from "i18next";
