+++ src/data/mystery-encounters/utils/encounter-phase-utils.ts
@@ -52,7 +52,7 @@ import { showEncounterText } from "#mystery-encounters/encounter-dialogue-utils"
 import type MysteryEncounterOption from "#mystery-encounters/mystery-encounter-option";
 import type PokemonData from "#system/pokemon-data";
 import { allTrainerConfigs } from "#trainer-configs/all-trainer-configs";
-import type { HeldModifierConfig } from "#types/held-modifier-config";
+import type { HeldModifierConfig } from "#types/modifiers-types";
 import type { PokemonSelectFilter } from "#types/pokemon-select-filter";
 import type { OptionSelectItem, OptionSelectModeConfig } from "#ui/option-select-config";
 import type { OptionSelectUiHandler } from "#ui/option-select-ui-handler";
